Implementa Swagger con SOLID
