+++ src/main/java/com/brokis/banco/modelo/Cuenta.java
@@ -10,6 +10,7 @@ import javax.persistence.*;
 public class Cuenta {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Deprecated
     private Long numeroCuenta;
     @Column(nullable = false)
     private String nombreTitular;
@@ -17,6 +18,7 @@ public class Cuenta {
     private int documentoTitular;
     @Column(nullable = false)
     private String fechaCreacion;
+    @Deprecated
     private double saldo = 0;
