+++ src/main/java/com/brokis/banco/repositorio/RepCuenta.java
@@ -1,4 +1,4 @@
-package com.brokis.banco.repocitorios;
+package com.brokis.banco.repositorio;
 
 import com.brokis.banco.modelo.Cuenta;
 import org.springframework.data.jpa.repository.JpaRepository;
