+++ src/main/java/com/brokis/banco/servicio/CreadorCuenta.java
@@ -1,0 +1,8 @@
+package com.brokis.banco.servicio;
+
+import com.brokis.banco.modelo.Cuenta;
+
+public interface CreadorCuenta {
+
+    Cuenta crearCuenta(Cuenta cuenta);
+}
