+++ src/main/java/com/brokis/banco/servicio/ServicioCuenta.java
@@ -1,9 +1,9 @@
-package com.brokis.banco.servicios;
+package com.brokis.banco.servicio;
 import com.brokis.banco.modelo.Cuenta;
 import java.util.Optional;
 
 public interface ServicioCuenta {
-    Cuenta crearCuenta(Cuenta cuenta);
+
     Optional<Cuenta> consultarSaldo(Long numeroCuneta); //El optional es para que no se rompa si no encuentra ninguna cuenta
     Cuenta realizarDeposito(Long numeroCuenta, double monto);
